feat: rethink the standalone API
